Moving required tests
